api-requirements.txt
onnxruntime-gpu
librosa
numpy
aiofiles
pydantic
python-multipart
jsonify
torchaudio